+++ himalaya/backend/numpy.py
@@ -2,12 +2,16 @@
 
 To use this backend, call ``himalaya.backend.set_backend("numpy")``.
 """
+
 import numpy as np
+
 try:
     import scipy.linalg as linalg
+
     use_scipy = True
 except ImportError:
     import numpy.linalg as linalg
+
     use_scipy = False
 
 ###############################################################################
@@ -36,16 +40,18 @@ def std_float64(array, axis=None, demean=True, keepdims=False):
     """Compute the standard deviation of X with double precision,
     and cast back the result to original dtype.
     """
-    return array.std(axis, dtype=np.float64,
-                     keepdims=keepdims).astype(array.dtype, copy=False)
+    return array.std(axis, dtype=np.float64, keepdims=keepdims).astype(
+        array.dtype, copy=False
+    )
 
 
 def mean_float64(array, axis=None, keepdims=False):
     """Compute the mean of X with double precision,
     and cast back the result to original dtype.
     """
-    return array.mean(axis, dtype=np.float64,
-                      keepdims=keepdims).astype(array.dtype, copy=False)
+    return array.mean(axis, dtype=np.float64, keepdims=keepdims).astype(
+        array.dtype, copy=False
+    )
 
 
 ###############################################################################
@@ -190,6 +196,7 @@ def asarray(a, dtype=None, order=None, device=None):
     # works from cupy
     try:
         import cupy
+
         return np.asarray(cupy.asnumpy(a), dtype=dtype, order=order)
     except Exception:
         pass
